cmms_alm-main/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.1.0
Django==5.1.3
django-cors-headers==4.4.0
django-countries==7.6.1
django-debug-toolbar==4.3.0
django-decouple==2.1
django-environ==0.11.2
django-filter==24.3
django-htmx==1.23.0
django-jazzmin==3.0.1
django-mail-templated==2.6.5
django-sslserver-v2==1.0
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.7
environ==1.0
environs==11.2.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
marshmallow==3.23.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.2
setuptools>=65.0.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.9.0
django-modeltranslation==0.19.15
